requirements.txt
langchain
langgraph
langchain-google-genai
langchain_openai
langchain-tavily
langmem
python-dotenv
streamlit
gradio
langchain_community
langchain_chroma
langchain_core
langchain_text_splitters
pypdf
langchain_experimental
lark
langchain-mcp-adapters
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
numpy
langchain-experimental
mcp
typing_extensions
openai
beautifulsoup4